[UPDATE 03/12/2025]: - Add base MOE implementation in blocks/mlp/moe.py - Add SoftMOE implementation in blocks/mlp/soft_moe.py - Minor bug fixed

+++ rectify/blocks/mamba/mamba.py
@@ -18,7 +18,7 @@ class MambaConfig:
     use_pscan: bool = True # use pscan for parallel computational
 
     def __post_init__(self):
-        self.d_inner = self.d_model * self.expand # inner dim
+        self.d_inner = int(self.d_model * self.expand) # inner dim
         if self.dt_rank == "auto":
             self.dt_rank = math.ceil(self.d_model / 16)
